+++ besu-plugins/linea-sequencer/acceptance-tests/src/test/java/linea/plugin/acc/test/LineaPluginTestBase.java
@@ -163,7 +163,8 @@ public abstract class LineaPluginTestBase extends AcceptanceTestBase {
                     "LineaTransactionPoolValidatorPlugin",
                     "LineaTransactionSelectorPlugin",
                     "LineaBundleEndpointsPlugin",
-                    "ForwardBundlesPlugin"));
+                    "ForwardBundlesPlugin",
+                    "LineaPermissioningPlugin"));
 
     return besu.create(nodeConfBuilder.build());
   }